Add qwen vl video worker
