+++ python/sglang/srt/multimodal/processors/qwen_vl.py
@@ -1,17 +1,11 @@
 import asyncio
-import math
-import os
 import re
 import time
 from typing import List, Union
 
-import numpy as np
 import torch
-import torchvision
 from PIL import Image
-from torchvision.transforms import InterpolationMode
 
-from sglang.srt.environ import envs
 from sglang.srt.layers.rotary_embedding import MRotaryEmbedding
 from sglang.srt.managers.schedule_batch import Modality, MultimodalDataItem
 from sglang.srt.models.qwen2_5_vl import Qwen2_5_VLForConditionalGeneration
@@ -23,125 +17,16 @@ from sglang.srt.multimodal.processors.base_processor import (
     BaseMultimodalProcessor as SGLangBaseProcessor,
 )
 from sglang.srt.multimodal.processors.base_processor import MultimodalSpecialTokens
+from sglang.srt.multimodal.processors.qwen_vl_video_worker import (
+    preprocess_video_task as qvl_preprocess_video_task,
+)
+from sglang.srt.multimodal.processors.qwen_vl_video_worker import (
+    resize_image_async as qvl_resize_image_async,
+)
 from sglang.srt.multimodal.processors.video_utils import VideoInput
 from sglang.utils import logger
 
 IMAGE_FACTOR = 28
-MIN_PIXELS = 4 * 28 * 28
-MAX_PIXELS = envs.SGLANG_IMAGE_MAX_PIXELS.get()
-MAX_RATIO = 200
-RESIZE_RESAMPLE = getattr(Image, envs.SGLANG_RESIZE_RESAMPLE.get(), None)
-if envs.SGLANG_RESIZE_RESAMPLE.is_set() and RESIZE_RESAMPLE is None:
-    logger.warning(
-        f"Invalid RESIZE_RESAMPLE value: '{envs.SGLANG_RESIZE_RESAMPLE.get()}'. "
-        f"Ignoring and using default."
-    )
-VIDEO_TOTAL_PIXELS = int(
-    float(os.environ.get("VIDEO_MAX_PIXELS", 128000 * 28 * 28 * 0.9))
-)
-
-VIDEO_MIN_PIXELS = 128 * 28 * 28
-VIDEO_MAX_PIXELS = 768 * 28 * 28
-FRAME_FACTOR = 2
-FPS = 2.0
-FPS_MIN_FRAMES = 4
-FPS_MAX_FRAMES = 768
-
-
-def smart_resize(
-    height: int,
-    width: int,
-    factor: int = IMAGE_FACTOR,
-    min_pixels: int = MIN_PIXELS,
-    max_pixels: int = MAX_PIXELS,
-) -> tuple[int, int]:
-    """
-    Rescales the image so that the following conditions are met:
-
-    1. Both dimensions (height and width) are divisible by 'factor'.
-
-    2. The total number of pixels is within the range ['min_pixels', 'max_pixels'].
-
-    3. The aspect ratio of the image is maintained as closely as possible.
-    """
-    if max(height, width) / min(height, width) > MAX_RATIO:
-        raise ValueError(
-            f"absolute aspect ratio must be smaller than {MAX_RATIO}, got {max(height, width) / min(height, width)}"
-        )
-    h_bar = max(factor, round_by_factor(height, factor))
-    w_bar = max(factor, round_by_factor(width, factor))
-    if h_bar * w_bar > max_pixels:
-        beta = math.sqrt((height * width) / max_pixels)
-        h_bar = floor_by_factor(height / beta, factor)
-        w_bar = floor_by_factor(width / beta, factor)
-    elif h_bar * w_bar < min_pixels:
-        beta = math.sqrt(min_pixels / (height * width))
-        h_bar = ceil_by_factor(height * beta, factor)
-        w_bar = ceil_by_factor(width * beta, factor)
-    return h_bar, w_bar
-
-
-def round_by_factor(number: int, factor: int) -> int:
-    """Returns the closest integer to 'number' that is divisible by 'factor'."""
-    return round(number / factor) * factor
-
-
-def ceil_by_factor(number: int, factor: int) -> int:
-    """Returns the smallest integer greater than or equal to 'number' that is divisible by 'factor'."""
-    return math.ceil(number / factor) * factor
-
-
-def floor_by_factor(number: int, factor: int) -> int:
-    """Returns the largest integer less than or equal to 'number' that is divisible by 'factor'."""
-    return math.floor(number / factor) * factor
-
-
-def smart_nframes(
-    ele: dict,
-    total_frames: int,
-    video_fps: int | float,
-) -> int:
-    """calculate the number of frames for video used for model inputs.
-
-    Args:
-        ele (dict): a dict contains the configuration of video.
-            support either `fps` or `nframes`:
-                - nframes: the number of frames to extract for model inputs.
-                - fps: the fps to extract frames for model inputs.
-                    - min_frames: the minimum number of frames of the video, only used when fps is provided.
-                    - max_frames: the maximum number of frames of the video, only used when fps is provided.
-        total_frames (int): the original total number of frames of the video.
-        video_fps (int | float): the original fps of the video.
-
-    Raises:
-        ValueError: nframes should in interval [FRAME_FACTOR, total_frames].
-
-    Returns:
-        int: the number of frames for video used for model inputs.
-    """
-    assert not (
-        "fps" in ele and "nframes" in ele
-    ), "Only accept either `fps` or `nframes`"
-    if "nframes" in ele:
-        nframes = round_by_factor(ele["nframes"], FRAME_FACTOR)
-    else:
-        fps = ele.get("fps", FPS)
-        min_frames = ceil_by_factor(ele.get("min_frames", FPS_MIN_FRAMES), FRAME_FACTOR)
-        max_frames = floor_by_factor(
-            ele.get("max_frames", min(FPS_MAX_FRAMES, total_frames)), FRAME_FACTOR
-        )
-        nframes = total_frames / video_fps * fps
-        if nframes > total_frames:
-            logger.warning(
-                f"smart_nframes: nframes[{nframes}] > total_frames[{total_frames}]"
-            )
-        nframes = min(min(max(nframes, min_frames), max_frames), total_frames)
-        nframes = floor_by_factor(nframes, FRAME_FACTOR)
-    if not (FRAME_FACTOR <= nframes and nframes <= total_frames):
-        raise ValueError(
-            f"nframes should in interval [{FRAME_FACTOR}, {total_frames}], but got {nframes}."
-        )
-    return nframes
 
 
 # Compatible with Qwen-VL & Qwen-Omni Series
@@ -230,96 +115,15 @@ class QwenVLImageProcessor(SGLangBaseProcessor):
             loop = asyncio.get_running_loop()
             return await loop.run_in_executor(
                 self.video_executor,
-                QwenVLImageProcessor.preprocess_video_task,
+                qvl_preprocess_video_task,
                 video,
                 image_factor,
             )
         else:
-            return self.preprocess_video_task(
+            return qvl_preprocess_video_task(
                 video,
                 image_factor,
             )
-
-    # process video, qwen-specific
-    @staticmethod
-    def preprocess_video_task(
-        video_input: VideoInput,
-        image_factor: int = IMAGE_FACTOR,
-        # vr: VideoReader, image_factor: int = IMAGE_FACTOR
-    ) -> torch.Tensor:
-        entry_time = time.perf_counter()
-        import torch
-        from decord import VideoReader, cpu, gpu
-        try:
-            from decord.bridge import decord_bridge
-            ctx = gpu(0)
-            _ = decord_bridge.get_ctx_device(ctx)
-        except Exception:
-            ctx = cpu(0)
-
-        vr = VideoReader(video_input.path, ctx=ctx)
-
-        ele = {}
-        total_frames, video_fps = len(vr), vr.get_avg_fps()
-        nframes = smart_nframes({}, total_frames=total_frames, video_fps=video_fps)
-        idx = np.linspace(0, total_frames - 1, num=nframes, dtype=np.int64)
-        idx = np.unique(idx)
-        video_np = vr.get_batch(idx).asnumpy()
-        video = torch.from_numpy(video_np).pin_memory()
-        video = video.permute(0, 3, 1, 2)  # Convert to TCHW format
-        nframes, _, height, width = video.shape
-        min_pixels = ele.get("min_pixels", VIDEO_MIN_PIXELS)
-        total_pixels = ele.get("total_pixels", VIDEO_TOTAL_PIXELS)
-        max_pixels = max(
-            min(VIDEO_MAX_PIXELS, total_pixels / nframes * FRAME_FACTOR),
-            int(min_pixels * 1.05),
-        )
-
-        get_batch_time = time.perf_counter()
-
-        max_pixels_supposed = ele.get("max_pixels", max_pixels)
-        if max_pixels_supposed > max_pixels:
-            logger.warning(
-                f"The given max_pixels[{max_pixels_supposed}] exceeds limit[{max_pixels}]."
-            )
-        max_pixels = min(max_pixels_supposed, max_pixels)
-        if "resized_height" in ele and "resized_width" in ele:
-            resized_height, resized_width = smart_resize(
-                ele["resized_height"],
-                ele["resized_width"],
-                factor=image_factor,
-            )
-        else:
-            resized_height, resized_width = smart_resize(
-                height,
-                width,
-                factor=image_factor,
-                min_pixels=min_pixels,
-                max_pixels=max_pixels,
-            )
-        smart_resize_time = time.perf_counter()
-        video = torchvision.transforms.functional.resize(
-            video,
-            [resized_height, resized_width],
-            interpolation=InterpolationMode.BILINEAR,
-        )
-        video = video.pin_memory()
-        video_metadata = {
-            "fps": video_fps,
-            "duration": total_frames / video_fps,
-            "total_num_frames": total_frames,
-            "frames_indices": idx,
-            "video_backend": "torchvision",
-        }
-        torchvision_resize_time = time.perf_counter()
-        logger.debug(
-            f"[preprocess_video Perf], "
-            f"get_batch_time: {(get_batch_time - entry_time) * 1000:.2f} ms, "
-            f"smart_resize_time: {(smart_resize_time - get_batch_time) * 1000:.2f} ms, "
-            f"torchvision_resize_time: {(torchvision_resize_time - smart_resize_time) * 1000:.2f} ms, "
-            f"total_time: {(torchvision_resize_time - entry_time) * 1000:.2f} ms"
-        )
-        return video, video_metadata
 
     async def process_mm_data_async(
         self,
@@ -340,12 +144,18 @@ class QwenVLImageProcessor(SGLangBaseProcessor):
         load_time = time.perf_counter()
         rid = getattr(request_obj, "rid", "anonymous_rid")
 
+        # Qwen-specific: resize images if they are raw Image objects
+        if base_output.images and isinstance(base_output.images[0], Image.Image):
+            resize_tasks = [
+                qvl_resize_image_async(image) for image in base_output.images
+            ]
+            base_output.images = await asyncio.gather(*resize_tasks)
+
         video_metadata = None
         # Now base_output.videos are List[VideoInput]
         if base_output.videos:
             videos_processed = [
-                await preprocess_video(video, video_config=self.video_config)
-                for video in base_output.videos
+                await self.preprocess_video(video) for video in base_output.videos
             ]
             base_output.videos, video_metadata = map(list, zip(*videos_processed))
